test: tighten toolbar priority assertions
