+++ Packages/iOS/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceTitleMenu.swift
@@ -1,4 +1,3 @@
-import CmuxMobileSupport
 import SwiftUI
 
 struct WorkspaceTitleMenu<Label: View, MenuContent: View>: View {
@@ -33,62 +32,5 @@ struct WorkspaceTitleMenu<Label: View, MenuContent: View>: View {
             .fixedSize(horizontal: false, vertical: true)
             .frame(minWidth: 0, maxWidth: .infinity, alignment: .center)
             .clipped()
-    }
-}
-
-enum MobileToolbarItemLayoutRole {
-    case compressibleTitle
-    case fixedTrailingControls
-
-    var swiftUILayoutPriority: Double {
-        switch self {
-        case .compressibleTitle:
-            return -1
-        case .fixedTrailingControls:
-            return 1
-        }
-    }
-}
-
-struct MobileToolbarPriorityHost<Content: View>: View {
-    let role: MobileToolbarItemLayoutRole
-    let content: Content
-
-    init(role: MobileToolbarItemLayoutRole, @ViewBuilder content: () -> Content) {
-        self.role = role
-        self.content = content()
-    }
-
-    @ViewBuilder
-    var body: some View {
-        content
-            .layoutPriority(role.swiftUILayoutPriority)
-    }
-}
-
-struct MobileWorkspacePriorityToolbar<Back: View, Title: View, Trailing: View>: View {
-    @ViewBuilder let back: () -> Back
-    @ViewBuilder let title: () -> Title
-    @ViewBuilder let trailing: () -> Trailing
-
-    var body: some View {
-        HStack(spacing: 8) {
-            back()
-                .layoutPriority(MobileToolbarItemLayoutRole.fixedTrailingControls.swiftUILayoutPriority)
-                .fixedSize()
-                .mobileGlassCompactToolbarControl()
-
-            title()
-                .layoutPriority(MobileToolbarItemLayoutRole.compressibleTitle.swiftUILayoutPriority)
-                .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                .mobileGlassCompactToolbarControl()
-                .clipped()
-
-            trailing()
-                .layoutPriority(MobileToolbarItemLayoutRole.fixedTrailingControls.swiftUILayoutPriority)
-                .fixedSize()
-                .mobileGlassCompactToolbarControl()
-        }
-        .frame(maxWidth: .infinity, minHeight: 44, maxHeight: 44, alignment: .center)
     }
 }
